2011-02-07 Roman - DAQ.System: The library libsensors disable is added for no allowed the library. - FIX:DAQ.System: Crash is fixed on exit for source "Hddtemp" exeptions at it destroy. - TArchives, Archive.FSArch: Multiscan for an archive directory is disabled at the program start. - !!:DB: User API functions fieldStruct(), fieldSeek(), fieldGet(), fieldSet() and fieldDel() for Table object is added. - !!DB.*: All warnings is resolved for "-Wall" compile option.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1257 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tvariant.cpp
@@ -459,7 +459,7 @@ TVariant XMLNodeObj::funcCall( const string &id, vector<TVariant> &prms )
     //  vl - value for attribute
     if( id == "setAttr" && prms.size() >= 2 )	{ propSet(prms[0].getS(),prms[1].getS()); return this; }
     // int childSize( ) - return childs counter for node
-    if( id == "childSize" )	return childSize();
+    if( id == "childSize" )	return (int)childSize();
     // XMLNodeObj childAdd(ElTp no = XMLNodeObj) - add node <no> as child to the node
     //  no - node object or name for new node
     if( id == "childAdd" )
